+++ LKXFrameworkForiOS/classes/Until(工具类)/LKXCommonFile/tool/LKXSpringCollectionViewFlowLayout.m
@@ -92,7 +92,7 @@
     {
         CGPoint anchorPoint = spring.anchorPoint;
         
-        CGFloat distanceFromTouch = fabsf(touchLocation.y - anchorPoint.y);
+        CGFloat distanceFromTouch = fabs(touchLocation.y - anchorPoint.y);
         CGFloat scrollResistance = distanceFromTouch / self.resistanceFactor;
         
         UICollectionViewLayoutAttributes *item = (UICollectionViewLayoutAttributes *)[spring.items firstObject];
